server/requirements.txt
python-socketio
aiohttp
eventlet